ref: guidance_tests
guidance_tests added to test select options error - extra -1000 in subprobs dictionary when dealing with a specific acronym test case - 2 with the same prefix and 1 with a different one. This leads to an np.log zero runtime warning. The probability of the existing tokens or the prefix with an incoming token is 0 due to the issue described earlier.